move relatedpods to the shared components
